+++ .github/workflows/test.yaml
@@ -335,6 +335,17 @@ jobs:
         run: |
           python -m pip install --upgrade pip
           pip install nox "uv<0.7.0"  # pin to unblock CI
+      - name: Run nox (Clickhouse Only)
+        env:
+          WEAVE_SENTRY_ENV: ci
+          CI: 1
+          WB_SERVER_HOST: http://wandbservice
+          WF_CLICKHOUSE_HOST: localhost
+          WEAVE_SERVER_DISABLE_ECOSYSTEM: 1
+          DD_TRACE_ENABLED: false
+        run: |
+          nox -e "tests-${{ matrix.python-version-major }}.${{ matrix.python-version-minor }}(shard='${{ matrix.nox-shard }}')" -- \
+            -m "not skip_clickhouse_client"
       - name: Run nox (Sqlite Only)
         env:
           WEAVE_SENTRY_ENV: ci
@@ -350,19 +361,9 @@ jobs:
           OPENAI_API_KEY: ${{ secrets.OPENAI_API_KEY }}
           DD_TRACE_ENABLED: false
         run: |
-          nox -e "tests-${{ matrix.python-version-major }}.${{ matrix.python-version-minor }}(shard='${{ matrix.nox-shard }}')"
-      - name: Run nox (Clickhouse Only)
-        env:
-          WEAVE_SENTRY_ENV: ci
-          CI: 1
-          WB_SERVER_HOST: http://wandbservice
-          WF_CLICKHOUSE_HOST: localhost
-          WEAVE_SERVER_DISABLE_ECOSYSTEM: 1
-          DD_TRACE_ENABLED: false
-        run: |
           nox -e "tests-${{ matrix.python-version-major }}.${{ matrix.python-version-minor }}(shard='${{ matrix.nox-shard }}')" -- \
-            -m "weave_client and not skip_clickhouse_client" \
-            --weave-server=clickhouse
+            -m "weave_client"
+            --trace-server=sqlite
   trace-tests-matrix-check: # This job does nothing and is only used for the branch protection
     if: always()
