+++ AGENTS.md
@@ -48,7 +48,7 @@ Focus on these primary test shards:
 
 ### Running Tests
 
-**IMPORTANT**: Any test depending on the `client` fixture runs agains either SQLite backend or Clickhouse. By default it will run againt SQLite for performance. However, it is critical to test both. Use the pytest custom flag `--weave-server=clickhouse` with `--clickhouse-process=true` to run tests against the clickhouse implementation.
+**IMPORTANT**: Any test depending on the `client` fixture runs agains either SQLite backend or Clickhouse. By default it will run againt SQLite for performance. However, it is critical to test both. Use the pytest custom flag `--trace-server=clickhouse` with `--clickhouse-process=true` to run tests against the clickhouse implementation.
 
 1. Run all tests in a specific shard: `nox --no-install -e "tests-3.12(shard='trace')"`
 2. Run a specific test by appending `-- [test]` like so: `nox --no-install -e "tests-3.12(shard='trace')" -- trace/test_client_trace.py::test_simple_op`
@@ -56,7 +56,7 @@ Focus on these primary test shards:
 
 *Important:* Since you don't have internet access, you must run `nox` with `--no-install`. We have pre-installed the requirements on the above shards.
 
-Therefore, a basic text command would look like: `nox --no-install -e "tests-3.12(shard='trace')" -- trace/test_client_trace.py::test_simple_op --weave-server=clickhouse --clickhouse-process=true`
+Therefore, a basic text command would look like: `nox --no-install -e "tests-3.12(shard='trace')" -- trace/test_client_trace.py::test_simple_op --trace-server=clickhouse --clickhouse-process=true`
 
 ## Typescript Testing Guidelines
 TODO: need to fill this out
