+++ tests/conftest.py
@@ -1,4 +1,3 @@
-import base64
 import contextlib
 import logging
 import os
@@ -11,16 +10,10 @@ from fastapi import FastAPI
 from fastapi.testclient import TestClient
 
 import weave
-from tests.conftest_lib.clickhouse_server import ensure_clickhouse_db
 from tests.trace.util import DummyTestException
+from tests.trace_server.conftest import TEST_ENTITY, get_trace_server_flag
 from weave.trace import autopatch, weave_client, weave_init
 from weave.trace.context.call_context import set_call_stack
-from weave.trace_server import (
-    clickhouse_trace_server_batched,
-    external_to_internal_trace_server_adapter,
-    sqlite_trace_server,
-)
-from weave.trace_server import environment as ts_env
 from weave.trace_server import trace_server_interface as tsi
 from weave.trace_server_bindings import remote_http_trace_server
 from weave.trace_server_bindings.caching_middleware_trace_server import (
@@ -91,27 +84,6 @@ def disable_datadog():
         os.environ["DD_TRACE_ENABLED"] = original_dd_trace
     elif "DD_TRACE_ENABLED" in os.environ:
         del os.environ["DD_TRACE_ENABLED"]
-
-
-def pytest_addoption(parser):
-    parser.addoption(
-        "--weave-server",
-        action="store",
-        default="sqlite",
-        help="Specify the client object to use: sqlite or clickhouse",
-    )
-    parser.addoption(
-        "--ch",
-        "--clickhouse",
-        action="store_true",
-        help="Use clickhouse server (shorthand for --weave-server=clickhouse)",
-    )
-    parser.addoption(
-        "--clickhouse-process",
-        action="store",
-        default="false",
-        help="Use a clickhouse process instead of a container",
-    )
 
 
 def pytest_collection_modifyitems(config, items):
@@ -218,124 +190,6 @@ def client_with_throwing_server(client):
         yield client
     finally:
         client.server = curr_server
-
-
-class TwoWayMapping:
-    def __init__(self):
-        self._ext_to_int_map = {}
-        self._int_to_ext_map = {}
-
-        # Useful for testing to ensure caching is working
-        self.stats = {
-            "ext_to_int": {
-                "hits": 0,
-                "misses": 0,
-            },
-            "int_to_ext": {
-                "hits": 0,
-                "misses": 0,
-            },
-        }
-
-    def ext_to_int(self, key, default=None):
-        if key not in self._ext_to_int_map:
-            if default is None:
-                raise ValueError(f"Key {key} not found")
-            if default in self._int_to_ext_map:
-                raise ValueError(f"Default {default} already in use")
-            self._ext_to_int_map[key] = default
-            self._int_to_ext_map[default] = key
-            self.stats["ext_to_int"]["misses"] += 1
-        else:
-            self.stats["ext_to_int"]["hits"] += 1
-        return self._ext_to_int_map[key]
-
-    def int_to_ext(self, key, default):
-        if key not in self._int_to_ext_map:
-            if default is None:
-                raise ValueError(f"Key {key} not found")
-            if default in self._ext_to_int_map:
-                raise ValueError(f"Default {default} already in use")
-            self._int_to_ext_map[key] = default
-            self._ext_to_int_map[default] = key
-            self.stats["int_to_ext"]["misses"] += 1
-        else:
-            self.stats["int_to_ext"]["hits"] += 1
-        return self._int_to_ext_map[key]
-
-
-def b64(s: str) -> str:
-    # Base64 encode the string
-    return base64.b64encode(s.encode("ascii")).decode("ascii")
-
-
-class DummyIdConverter(external_to_internal_trace_server_adapter.IdConverter):
-    def __init__(self):
-        self._project_map = TwoWayMapping()
-        self._run_map = TwoWayMapping()
-        self._user_map = TwoWayMapping()
-
-    def ext_to_int_project_id(self, project_id: str) -> str:
-        return self._project_map.ext_to_int(project_id, b64(project_id))
-
-    def int_to_ext_project_id(self, project_id: str) -> typing.Optional[str]:
-        return self._project_map.int_to_ext(project_id, b64(project_id))
-
-    def ext_to_int_run_id(self, run_id: str) -> str:
-        return self._run_map.ext_to_int(run_id, b64(run_id) + ":" + run_id)
-
-    def int_to_ext_run_id(self, run_id: str) -> str:
-        exp = run_id.split(":")[1]
-        return self._run_map.int_to_ext(run_id, exp)
-
-    def ext_to_int_user_id(self, user_id: str) -> str:
-        return self._user_map.ext_to_int(user_id, b64(user_id))
-
-    def int_to_ext_user_id(self, user_id: str) -> str:
-        return self._user_map.int_to_ext(user_id, b64(user_id))
-
-
-class TestOnlyUserInjectingExternalTraceServer(
-    external_to_internal_trace_server_adapter.ExternalTraceServer
-):
-    def __init__(
-        self,
-        internal_trace_server: tsi.TraceServerInterface,
-        id_converter: external_to_internal_trace_server_adapter.IdConverter,
-        user_id: str,
-    ):
-        super().__init__(internal_trace_server, id_converter)
-        self._user_id = user_id
-
-    def call_start(self, req: tsi.CallStartReq) -> tsi.CallStartRes:
-        req.start.wb_user_id = self._user_id
-        return super().call_start(req)
-
-    def calls_delete(self, req: tsi.CallsDeleteReq) -> tsi.CallsDeleteRes:
-        req.wb_user_id = self._user_id
-        return super().calls_delete(req)
-
-    def call_update(self, req: tsi.CallUpdateReq) -> tsi.CallUpdateRes:
-        req.wb_user_id = self._user_id
-        return super().call_update(req)
-
-    def feedback_create(self, req: tsi.FeedbackCreateReq) -> tsi.FeedbackCreateRes:
-        req.wb_user_id = self._user_id
-        return super().feedback_create(req)
-
-    def cost_create(self, req: tsi.CostCreateReq) -> tsi.CostCreateRes:
-        req.wb_user_id = self._user_id
-        return super().cost_create(req)
-
-    def actions_execute_batch(
-        self, req: tsi.ActionsExecuteBatchReq
-    ) -> tsi.ActionsExecuteBatchRes:
-        req.wb_user_id = self._user_id
-        return super().actions_execute_batch(req)
-
-    def obj_create(self, req: tsi.ObjCreateReq) -> tsi.ObjCreateRes:
-        req.obj.wb_user_id = self._user_id
-        return super().obj_create(req)
 
 
 # https://docs.pytest.org/en/7.1.x/example/simple.html#pytest-current-test-environment-variable
@@ -498,62 +352,31 @@ def create_client(
     autopatch_settings: typing.Optional[autopatch.AutopatchSettings] = None,
     global_attributes: typing.Optional[dict[str, typing.Any]] = None,
 ) -> weave_init.InitializedClient:
-    inited_client = None
-    weave_server_flag = request.config.getoption("--weave-server")
-
-    # Check if -ch/--clickhouse flag is used
-    if request.config.getoption("--clickhouse"):
-        weave_server_flag = "clickhouse"
+    trace_server_flag = get_trace_server_flag(request)
+    if trace_server_flag == "prod":
+        # Note: this is only for local dev testing and should be removed
+        return weave_init.init_weave("dev_testing")
 
     server: tsi.TraceServerInterface
-    entity = "shawn"
-    project = "test-project"
 
-    if weave_server_flag == "sqlite":
-        sqlite_server = sqlite_trace_server.SqliteTraceServer(
-            "file::memory:?cache=shared"
-        )
-        sqlite_server.drop_tables()
-        sqlite_server.setup_tables()
-        server = TestOnlyUserInjectingExternalTraceServer(
-            sqlite_server, DummyIdConverter(), entity
-        )
-    elif weave_server_flag == "clickhouse":
-        assert ensure_clickhouse_db is not None
-        host, port = request.getfixturevalue("ensure_clickhouse_db")
+    if trace_server_flag == "sqlite" or trace_server_flag == "clickhouse":
+        server = request.getfixturevalue("trace_server")
+    elif trace_server_flag.startswith("http"):
+        server = remote_http_trace_server.RemoteHTTPTraceServer(trace_server_flag)
+    else:
+        raise ValueError(f"Invalid trace server: {trace_server_flag}")
 
-        ch_server = clickhouse_trace_server_batched.ClickHouseTraceServer(
-            host=host,
-            port=port,
-        )
-        ch_server.ch_client.command("DROP DATABASE IF EXISTS db_management")
-        ch_server.ch_client.command(
-            f"DROP DATABASE IF EXISTS {ts_env.wf_clickhouse_database()}"
-        )
-        ch_server._run_migrations()
-        server = TestOnlyUserInjectingExternalTraceServer(
-            ch_server, DummyIdConverter(), entity
-        )
-    elif weave_server_flag.startswith("http"):
-        remote_server = remote_http_trace_server.RemoteHTTPTraceServer(
-            weave_server_flag
-        )
-        server = remote_server
-    elif weave_server_flag == ("prod"):
-        inited_client = weave_init.init_weave("dev_testing")
-
-    if inited_client is None:
-        # Removing this as it lead to passing tests that were not passing in prod!
-        # Keeping off for now until it is the default behavior.
-        # os.environ["WEAVE_USE_SERVER_CACHE"] = "true"
-        server = CachingMiddlewareTraceServer.from_env(server)
-        client = TestOnlyFlushingWeaveClient(
-            entity, project, make_server_recorder(server)
-        )
-        inited_client = weave_init.InitializedClient(client)
-        autopatch.autopatch(autopatch_settings)
-        if global_attributes is not None:
-            weave.trace.api._global_attributes = global_attributes
+    # Removing this as it lead to passing tests that were not passing in prod!
+    # Keeping off for now until it is the default behavior.
+    # os.environ["WEAVE_USE_SERVER_CACHE"] = "true"
+    server = CachingMiddlewareTraceServer.from_env(server)
+    client = TestOnlyFlushingWeaveClient(
+        TEST_ENTITY, "test-project", make_server_recorder(server)
+    )
+    inited_client = weave_init.InitializedClient(client)
+    autopatch.autopatch(autopatch_settings)
+    if global_attributes is not None:
+        weave.trace.api._global_attributes = global_attributes
 
     return inited_client
