+++ tests/trace_server/conftest_lib/trace_server_external_adapter.py
@@ -1,0 +1,146 @@
+import base64
+import typing
+from weave.trace_server import (
+
+    external_to_internal_trace_server_adapter,
+
+)
+
+from weave.trace_server import trace_server_interface as tsi
+
+class TwoWayMapping:
+    def __init__(self):
+        self._ext_to_int_map = {}
+        self._int_to_ext_map = {}
+
+        # Useful for testing to ensure caching is working
+        self.stats = {
+            "ext_to_int": {
+                "hits": 0,
+                "misses": 0,
+            },
+            "int_to_ext": {
+                "hits": 0,
+                "misses": 0,
+            },
+        }
+
+    def ext_to_int(self, key, default=None):
+        if key not in self._ext_to_int_map:
+            if default is None:
+                raise ValueError(f"Key {key} not found")
+            if default in self._int_to_ext_map:
+                raise ValueError(f"Default {default} already in use")
+            self._ext_to_int_map[key] = default
+            self._int_to_ext_map[default] = key
+            self.stats["ext_to_int"]["misses"] += 1
+        else:
+            self.stats["ext_to_int"]["hits"] += 1
+        return self._ext_to_int_map[key]
+
+    def int_to_ext(self, key, default):
+        if key not in self._int_to_ext_map:
+            if default is None:
+                raise ValueError(f"Key {key} not found")
+            if default in self._ext_to_int_map:
+                raise ValueError(f"Default {default} already in use")
+            self._int_to_ext_map[key] = default
+            self._ext_to_int_map[default] = key
+            self.stats["int_to_ext"]["misses"] += 1
+        else:
+            self.stats["int_to_ext"]["hits"] += 1
+        return self._int_to_ext_map[key]
+
+
+def b64(s: str) -> str:
+    # Base64 encode the string
+    return base64.b64encode(s.encode("ascii")).decode("ascii")
+
+
+class DummyIdConverter(external_to_internal_trace_server_adapter.IdConverter):
+    def __init__(self):
+        self._project_map = TwoWayMapping()
+        self._run_map = TwoWayMapping()
+        self._user_map = TwoWayMapping()
+
+    def ext_to_int_project_id(self, project_id: str) -> str:
+        return self._project_map.ext_to_int(project_id, b64(project_id))
+
+    def int_to_ext_project_id(self, project_id: str) -> typing.Optional[str]:
+        return self._project_map.int_to_ext(project_id, b64(project_id))
+
+    def ext_to_int_run_id(self, run_id: str) -> str:
+        return self._run_map.ext_to_int(run_id, b64(run_id) + ":" + run_id)
+
+    def int_to_ext_run_id(self, run_id: str) -> str:
+        exp = run_id.split(":")[1]
+        return self._run_map.int_to_ext(run_id, exp)
+
+    def ext_to_int_user_id(self, user_id: str) -> str:
+        return self._user_map.ext_to_int(user_id, b64(user_id))
+
+    def int_to_ext_user_id(self, user_id: str) -> str:
+        return self._user_map.int_to_ext(user_id, b64(user_id))
+
+
+class TestOnlyUserInjectingExternalTraceServer(
+    external_to_internal_trace_server_adapter.ExternalTraceServer
+):
+    def __init__(
+        self,
+        internal_trace_server: tsi.TraceServerInterface,
+        id_converter: external_to_internal_trace_server_adapter.IdConverter,
+        user_id: str,
+    ):
+        super().__init__(internal_trace_server, id_converter)
+        self._user_id = user_id
+
+    def call_start(self, req: tsi.CallStartReq) -> tsi.CallStartRes:
+        req.start.wb_user_id = self._user_id
+        return super().call_start(req)
+
+    def calls_delete(self, req: tsi.CallsDeleteReq) -> tsi.CallsDeleteRes:
+        req.wb_user_id = self._user_id
+        return super().calls_delete(req)
+
+    def call_update(self, req: tsi.CallUpdateReq) -> tsi.CallUpdateRes:
+        req.wb_user_id = self._user_id
+        return super().call_update(req)
+
+    def feedback_create(self, req: tsi.FeedbackCreateReq) -> tsi.FeedbackCreateRes:
+        req.wb_user_id = self._user_id
+        return super().feedback_create(req)
+
+    def cost_create(self, req: tsi.CostCreateReq) -> tsi.CostCreateRes:
+        req.wb_user_id = self._user_id
+        return super().cost_create(req)
+
+    def actions_execute_batch(
+        self, req: tsi.ActionsExecuteBatchReq
+    ) -> tsi.ActionsExecuteBatchRes:
+        req.wb_user_id = self._user_id
+        return super().actions_execute_batch(req)
+
+    def obj_create(self, req: tsi.ObjCreateReq) -> tsi.ObjCreateRes:
+        req.obj.wb_user_id = self._user_id
+        return super().obj_create(req)
+    
+    def run_model(self, req: tsi.RunModelReq) -> tsi.RunModelRes:
+        req.wb_user_id = self._user_id
+        return super().run_model(req)
+    
+    def run_scorer(self, req: tsi.RunScorerReq) -> tsi.RunScorerRes:
+        req.wb_user_id = self._user_id
+        return super().run_scorer(req)
+    
+    def queue_evaluation(self, req: tsi.QueueEvaluationReq) -> tsi.QueueEvaluationRes:
+        req.wb_user_id = self._user_id
+        return super().queue_evaluation(req)
+    
+
+def externalize_trace_server(trace_server: tsi.TraceServerInterface, user_id: str = "test_user") -> tsi.TraceServerInterface:
+    return TestOnlyUserInjectingExternalTraceServer(
+        trace_server,
+        DummyIdConverter(),
+        user_id,
+    )
