+++ tests/trace_server/test_evaluation_execution_api.py
@@ -1,0 +1,2 @@
+def test_run_model(ch_trace_server):
+    assert ch_trace_server == 5
